Fix Async Retry Event Handling
